test is inprogress

+++ common/ifs-sil-resources/src/main/java/org/innovateuk/ifs/sil/crm/resource/SilIMApplicationLocationInfo.java
@@ -45,7 +45,7 @@ public class SilIMApplicationLocationInfo {
     private String fundingDecisionStatus;
 
     @JsonProperty("durationInMonths")
-    private Integer durationInMonths;
+    private Long durationInMonths;
 
     @JsonProperty("completion")
     @JsonSerialize(using = PercentageSerializer.class)
